WIP parse function now works as udf
