+++ optimus/dataframe/extension.py
@@ -13,8 +13,7 @@ from pyspark.sql import functions as F
 from pyspark.sql.types import *
 
 from optimus.bumblebee import Comm
-from optimus.helpers.functions import val_to_list
-from optimus.helpers.check import is_str
+from infer import is_str
 from optimus.helpers.columns import parse_columns
 from optimus.helpers.constants import RELATIVE_ERROR
 from optimus.helpers.decorators import *
