+++ optimus/dataframe/rows.py
@@ -8,9 +8,10 @@ from pyspark.sql import functions as F
 # Helpers
 import optimus as op
 from optimus.audf import filter_row_by_data_type as fbdt
-from optimus.helpers.check import is_list_of_str_or_int, is_list_of_tuples, is_list_of_dataframes, is_dataframe
+from infer import is_list_of_str_or_int, is_list_of_dataframes, is_list_of_tuples, is_dataframe, \
+    PYSPARK_NUMERIC_TYPES
 from optimus.helpers.columns import parse_columns, validate_columns_names
-from optimus.helpers.constants import Actions, PYSPARK_NUMERIC_TYPES
+from optimus.helpers.constants import Actions
 from optimus.helpers.converter import one_list_to_val
 from optimus.helpers.decorators import add_attr
 from optimus.helpers.functions import append as append_df
