+++ optimus/helpers/columns_expression.py
@@ -3,8 +3,8 @@ import itertools
 
 from pyspark.sql import functions as F
 
-from optimus.helpers.check import is_column_a, is_numeric
-from optimus.helpers.constants import PYSPARK_NUMERIC_TYPES, PYSPARK_STRING_TYPES
+from infer import is_numeric, PYSPARK_NUMERIC_TYPES, PYSPARK_STRING_TYPES
+from optimus.helpers.check import is_column_a
 from optimus.helpers.converter import val_to_list
 from optimus.helpers.functions import create_buckets
