+++ optimus/helpers/parser.py
@@ -1,6 +1,5 @@
-from optimus.helpers.constants import SPARK_DTYPES_DICT, SPARK_SHORT_DTYPES, PYTHON_SHORT_TYPES, \
-    SPARK_DTYPES_DICT_OBJECTS
 from optimus.helpers.logger import logger
+from infer import PYTHON_SHORT_TYPES, SPARK_SHORT_DTYPES, SPARK_DTYPES_DICT
 
 
 def compress_dict(lst, col_name):
@@ -18,8 +17,18 @@ def compress_list(lst):
     return _result
 
 
+def parse_python_dtypes(value):
+    """
+    Get a spark data type from a string
+    :param value:
+    :return:
+    """
+    return PYTHON_SHORT_TYPES[value.lower()]
+
+
 def parse_col_names_funcs_to_keys(data):
-    from optimus.helpers.check import is_numeric, is_nan
+    from infer import is_numeric
+    from infer import is_nan
 
     """
     Helper function that return a formatted json with function:value inside columns. Transform from
@@ -75,38 +84,6 @@ def parse_spark_dtypes(value):
     try:
         data_type = [SPARK_DTYPES_DICT[SPARK_SHORT_DTYPES[v]] for v in value]
     except KeyError:
-        data_type = value
-
-    if isinstance(data_type, list) and len(data_type) == 1:
-        result = data_type[0]
-    else:
-        result = data_type
-
-    return result
-
-
-def parse_python_dtypes(value):
-    """
-    Get a spark data type from a string
-    :param value:
-    :return:
-    """
-    return PYTHON_SHORT_TYPES[value.lower()]
-
-
-def parse_spark_class_dtypes(value):
-    """
-    Get a pyspark data class from a string data type representation. for example 'StringType()' from 'string'
-    :param value:
-    :return:
-    """
-    if not isinstance(value, list):
-        value = [value]
-
-    try:
-        data_type = [SPARK_DTYPES_DICT_OBJECTS[SPARK_SHORT_DTYPES[v]] for v in value]
-
-    except (KeyError, TypeError):
         data_type = value
 
     if isinstance(data_type, list) and len(data_type) == 1:
