+++ optimus/io/load.py
@@ -8,7 +8,7 @@ from urllib.request import Request, urlopen
 import pandas as pd
 from packaging import version
 
-from optimus.helpers.check import is_url
+from infer import is_url
 from optimus.helpers.columns import replace_columns_special_characters
 from optimus.helpers.logger import logger
 from optimus.spark import Spark
