+++ optimus/ml/models.py
@@ -4,7 +4,7 @@ from pyspark.sql import functions as F
 from pysparkling import *
 from pysparkling.ml import H2OAutoML, H2ODeepLearning, H2OXGBoost, H2OGBM
 
-from optimus.helpers.check import is_dataframe, is_str
+from infer import is_str, is_dataframe
 from optimus.helpers.columns import parse_columns, name_col
 from optimus.ml.encoding import string_to_index, vector_assembler
 from optimus.spark import Spark
