+++ optimus/outliers/modified_z_score.py
@@ -1,6 +1,6 @@
 from pyspark.sql import functions as F
 
-from optimus.helpers.check import is_dataframe, is_numeric
+from infer import is_numeric, is_dataframe
 from optimus.helpers.columns import parse_columns, name_col
 from optimus.helpers.constants import RELATIVE_ERROR
 from optimus.helpers.converter import one_list_to_val
