+++ tests/test_df_cols.py
@@ -71,10 +71,10 @@ class Test_df_cols(unittest.TestCase):
 		actual_df =source_df.cols.count_by_dtypes('*',infer=True)
 		expected_value ={'names': {'string': 6, 'null': 1, 'int': 0, 'decimal': 0, 'boolean': 0, 'date': 0, 'array': 0, 'object': 0, 'gender': 0, 'ip': 0, 'url': 0, 'email': 0, 'credit_card_number': 0, 'zip_code': 0, 'missing': 0}, 'height(ft)': {'smallint': 5, 'null': 2}, 'function': {'string': 6, 'null': 1, 'int': 0, 'decimal': 0, 'boolean': 0, 'date': 0, 'array': 0, 'object': 0, 'gender': 0, 'ip': 0, 'url': 0, 'email': 0, 'credit_card_number': 0, 'zip_code': 0, 'missing': 0}, 'rank': {'tinyint': 6, 'null': 1}, 'age': {'int': 6, 'null': 1}, 'weight(t)': {'float': 5, 'null': 2}, 'japanese name': {'array': 6, 'null': 1}, 'last position seen': {'array': 3, 'date': 1, 'null': 3, 'int': 0, 'decimal': 0, 'string': 0, 'boolean': 0, 'object': 0, 'gender': 0, 'ip': 0, 'url': 0, 'email': 0, 'credit_card_number': 0, 'zip_code': 0, 'missing': 0}, 'date arrival': {'date': 6, 'null': 1, 'int': 0, 'decimal': 0, 'string': 0, 'boolean': 0, 'array': 0, 'object': 0, 'gender': 0, 'ip': 0, 'url': 0, 'email': 0, 'credit_card_number': 0, 'zip_code': 0, 'missing': 0}, 'last date seen': {'date': 6, 'null': 1, 'int': 0, 'decimal': 0, 'string': 0, 'boolean': 0, 'array': 0, 'object': 0, 'gender': 0, 'ip': 0, 'url': 0, 'email': 0, 'credit_card_number': 0, 'zip_code': 0, 'missing': 0}, 'attributes': {'array': 6, 'null': 1}, 'Date Type': {'date': 6, 'null': 1}, 'timestamp': {'timestamp': 6, 'null': 1}, 'Cybertronian': {'boolean': 6, 'null': 1}, 'function(binary)': {'binary': 6, 'null': 1}, 'NullType': {'null': 7}}
 		self.assertDictEqual(deep_sort(expected_value),  deep_sort(actual_df))
-	def test_cols_count_by_dtypes_mismatch(self):
+	def test_cols_count_mismatch(self):
 		source_df=op.create.df([('names', StringType(), True),('height(ft)', IntegerType(), True),('function', StringType(), True),('rank', IntegerType(), True),('age', IntegerType(), True),('weight(t)', FloatType(), True),('japanese name', ArrayType(StringType(),True), True),('last position seen', StringType(), True),('date arrival', StringType(), True),('last date seen', StringType(), True),('attributes', ArrayType(FloatType(),True), True),('DateType', DateType(), True),('Timestamp', TimestampType(), True),('Cybertronian', BooleanType(), True),('function(binary)', BinaryType(), False),('NullType', NullType(), True)], [('31/12/2019', 28, '1978-12-20', 10, 5000000, 4.300000190734863, ['Inochi', 'Convoy'], '19.442735,-99.201111', '1980/04/10', '2016/09/10', [8.53439998626709, 4300.0], datetime.date(2016, 9, 10), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'Leader'), None), ('bumbl#ebéé  ', 17, 'Espionage', 7, 5000000, 2.0, ['Bumble', 'Goldback'], '10.642707,-71.612534', '1980/04/10', '2015/08/10', [5.334000110626221, 2000.0], datetime.date(2015, 8, 10), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'Espionage'), None), ('ironhide&', 26, 'Security', 7, 5000000, 4.0, ['Roadbuster'], '37.789563,-122.400356', '1980/04/10', '2014/07/10', [7.924799919128418, 4000.0], datetime.date(2014, 6, 24), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'Security'), None), ('Jazz', 13, 'First Lieutenant', 8, 5000000, 1.7999999523162842, ['Meister'], '33.670666,-117.841553', '1980/04/10', '2013/06/10', [3.962399959564209, 1800.0], datetime.date(2013, 6, 24), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'First Lieutenant'), None), ('Megatron', None, 'None', 10, 5000000, 5.699999809265137, ['Megatron'], None, '1980/04/10', '2012/05/10', [None, 5700.0], datetime.date(2012, 5, 10), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'None'), None), ('Metroplex_)^$', 300, 'Battle Station', 8, 5000000, None, ['Metroflex'], None, '1980/04/10', '2011/04/10', [91.44000244140625, None], datetime.date(2011, 4, 10), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'Battle Station'), None), ('1', 2, '3', 4, 5, 6.0, ['7'], '8', '1980/04/10', '2011/04/10', [11.0], datetime.date(2011, 4, 10), datetime.datetime(2014, 6, 24, 0, 0), True, bytearray(b'15'), None)])
-		actual_df =source_df.cols.count_by_dtypes('*',infer=False,mismatch={'names': 'dd/mm/yyyy', 'height(ft)': '^([0-2][0-9]|(3)[0-1])(\\/)(((0)[0-9])|((1)[0-2]))(\\/)\\d{4}$', 'function': 'yyyy-mm-dd'})
-		expected_value ={'names': {'null': 0, 'missing': 0, 'string': 7, 'mismatch': 6}, 'height(ft)': {'null': 1, 'missing': 0, 'int': 6, 'mismatch': 0}, 'function': {'null': 0, 'missing': 0, 'string': 7, 'mismatch': 6}, 'rank': {'null': 0, 'missing': 0, 'int': 7, 'mismatch': 0}, 'age': {'null': 0, 'missing': 0, 'int': 7, 'mismatch': 0}, 'weight(t)': {'null': 1, 'missing': 0, 'decimal': 6, 'mismatch': 0}, 'japanese name': {'null': 0, 'missing': 0, 'array': 7, 'mismatch': 0}, 'last position seen': {'null': 2, 'missing': 0, 'string': 5, 'mismatch': 0}, 'date arrival': {'null': 0, 'missing': 0, 'string': 7, 'mismatch': 0}, 'last date seen': {'null': 0, 'missing': 0, 'string': 7, 'mismatch': 0}, 'attributes': {'null': 0, 'missing': 0, 'array': 7, 'mismatch': 0}, 'DateType': {'null': 0, 'missing': 0, 'date': 7, 'mismatch': 0}, 'Timestamp': {'null': 0, 'missing': 0, 'date': 7, 'mismatch': 0}, 'Cybertronian': {'null': 0, 'missing': 0, 'boolean': 7, 'mismatch': 0}, 'function(binary)': {'null': 0, 'missing': 0, 'binary': 7, 'mismatch': 0}, 'NullType': {'null': 7, 'missing': 0, 'mismatch': 0}}
+		actual_df =source_df.cols.count_mismatch(m)
+		expected_value ={'names': {'mismatch': 6, 'int': 1, 'null': 0, 'missing': 0}}
 		self.assertDictEqual(deep_sort(expected_value),  deep_sort(actual_df))
 	@staticmethod
 	def test_cols_count_na():
@@ -315,13 +315,13 @@ class Test_df_cols(unittest.TestCase):
 		actual_df =source_df.cols.mode('height(ft)')
 		actual_df =json_enconding(actual_df)
 		expected_value =json_enconding(None)
-		assert (expected_value == actual_df)
+		assert(expected_value == actual_df)
 	@staticmethod
 	def test_cols_mode_all_columns():
 		actual_df =source_df.cols.mode('*')
 		actual_df =json_enconding(actual_df)
 		expected_value =json_enconding([{'names': None}, {'height(ft)': None}, {'function': None}, {'rank': [8, 7, 10]}, {'age': 5000000}, {'weight(t)': None}, {'japanese name': None}, {'last position seen': None}, {'date arrival': '1980/04/10'}, {'last date seen': None}, {'attributes': None}, {'Date Type': None}, {'timestamp': datetime.datetime(2014, 6, 24, 0, 0)}, {'Cybertronian': True}, {'function(binary)': None}, {'NullType': None}])
-		assert (expected_value == actual_df)
+		assert(expected_value == actual_df)
 	@staticmethod
 	def test_cols_move_after():
 		actual_df =source_df.cols.move('rank','after','attributes')
